pyproject.toml
[project]
name = "server-management-lib"
version = "0.1.0"
description = "Shared library for server management MCPs. Provides SSH execution, security validation, config loading, and HTTP clients for InfluxDB/Prometheus."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "asyncssh>=2.14.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0.0",
    # Linting
    "ruff>=0.9.0",
    # Type checking
    "ty>=0.0.1a10",
    # Git hooks
    "pre-commit>=4.0.0",
    # Markdown linting
    "pymarkdownlnt>=0.9.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks integration tests that require real SSH connection",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["server_management_lib"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]  # Allow long lines in tests

[tool.ty.environment]
python-version = "3.11"

[tool.ty.rules]
unresolved-attribute = "ignore"
invalid-argument-type = "ignore"
